feat: reduce load time by refactoring Header, bouncing ai thinking, etc.
