+++ app/level/[id]/loading.tsx
@@ -2,7 +2,7 @@ import { Skeleton } from '@/components/custom/skeleton';
 
 export default function Loading() {
   return (
-    <section className='flex justify-center h-full pt-20 px-4'>
+    <section className='flex justify-center h-full w-full pt-20 px-4'>
       <Skeleton numberOfRows={10} />
     </section>
   );
