+++ app/roadmap/roadmap.md
@@ -6,6 +6,7 @@
 
 - Convenient way to allow topic generated from AI Mode to be contributed by providing logged in user the option to submit the level.
 - Logged in user is able to view his/her last result from Home, Menu or User Menu. The last result is restored from browser cache, therefore, logging out and log back in, or switch user, or clear browser cache, will remove the saved result and unable to view.
+- Greatly improve the performance of Taboo AI. Now it loads much faster.
 
 ### 13 Sep 2023 - V3.0.3
