+++ components/custom/back-button.tsx
@@ -1,39 +1,23 @@
 'use client';
 
-import { Tooltip, TooltipContent, TooltipTrigger } from '../ui/tooltip';
-import { ArrowLeft, Home } from 'lucide-react';
+import { ArrowLeft } from 'lucide-react';
 import { useRouter } from 'next/navigation';
 import IconButton from '../ui/icon-button';
 
 interface BackButtonProps {
-  href?: string;
-  customIcon?: React.ReactElement;
+  customBackHref?: string;
 }
 
-export function BackButton({ href, customIcon }: BackButtonProps) {
+export function BackButton({ customBackHref }: BackButtonProps) {
   const router = useRouter();
-  const back = () => {
-    if (href) {
-      router.push(href);
-    } else {
-      router.back();
-    }
-  };
-
   return (
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <IconButton
-          tooltip={href === '/' ? 'Go to Home' : 'Go Back'}
-          aria-label='Click to navigate back'
-          onClick={back}
-        >
-          {customIcon ?? (href === '/' ? <Home /> : <ArrowLeft />)}
-        </IconButton>
-      </TooltipTrigger>
-      <TooltipContent>
-        {`${href === '/' ? 'Go to home' : 'Back'}`}
-      </TooltipContent>
-    </Tooltip>
+    <IconButton
+      tooltip='Go Back'
+      onClick={() => {
+        customBackHref ? router.push(customBackHref) : router.back();
+      }}
+    >
+      <ArrowLeft />
+    </IconButton>
   );
 }
