+++ components/custom/menu-button.tsx
@@ -1,0 +1,18 @@
+'use client';
+
+import { CustomEventKey, EventManager } from '@/lib/event-manager';
+import { Menu } from 'lucide-react';
+import IconButton from '../ui/icon-button';
+
+export function MenuButton() {
+  return (
+    <IconButton
+      tooltip='Open Menu'
+      onClick={() => {
+        EventManager.fireEvent(CustomEventKey.TOGGLE_MENU, true);
+      }}
+    >
+      <Menu />
+    </IconButton>
+  );
+}
