+++ lib/hooks/useLocalStorage.tsx
@@ -4,11 +4,16 @@ import { useDebugValue, useEffect, useState } from 'react';
 
 export const useLocalStorage = <S,>(key: string) => {
   const [state, setState] = useState<S>();
+  const [loading, setLoading] = useState(true);
   useDebugValue(state);
 
   useEffect(() => {
     const item = localStorage.getItem(key);
     if (item) setState(parse(item));
+    setLoading(false);
+    return () => {
+      setLoading(true);
+    };
   }, []);
 
   const setItem = (s: S) => {
@@ -21,7 +26,7 @@ export const useLocalStorage = <S,>(key: string) => {
     setState(undefined);
   };
 
-  return { item: state, setItem, clearItem };
+  return { item: state, setItem, clearItem, loading };
 };
 
 const parse = (value: string) => {
