+++ next.config.js
@@ -62,4 +62,8 @@ const nextConfig = {
   },
 };
 
-module.exports = withPWA(nextConfig);
+const withBundleAnalyzer = require('@next/bundle-analyzer')({
+  enabled: process.env.ANALYZE === 'true',
+});
+
+module.exports = withBundleAnalyzer(withPWA(nextConfig));
