+++ package.json
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^3.3.1",
+    "@next/bundle-analyzer": "^13.4.19",
     "@radix-ui/react-accordion": "^1.1.2",
     "@radix-ui/react-alert-dialog": "^1.0.4",
     "@radix-ui/react-aspect-ratio": "^1.0.3",
@@ -53,13 +54,13 @@
     "lru-cache": "^7.12.0",
     "lucide-react": "^0.274.0",
     "moment": "^2.29.4",
-    "next": "^13.4.6",
+    "next": "^13.4.19",
     "next-pwa": "^5.6.0",
     "next-themes": "^0.2.1",
     "raw-loader": "^0.5.1",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-device-detect": "^2.2.2",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.46.1",
     "react-icons": "^4.10.1",
     "react-markdown": "^8.0.5",
